break out druid models
