+++ web-console/src/console-application.tsx
@@ -18,6 +18,7 @@
 
 import { Intent } from '@blueprintjs/core';
 import { IconNames } from '@blueprintjs/icons';
+import axios from 'axios';
 import classNames from 'classnames';
 import React from 'react';
 import { HashRouter, Route, Switch } from 'react-router-dom';
@@ -55,6 +56,10 @@ export class ConsoleApplication extends React.PureComponent<
   ConsoleApplicationState
 > {
   private capabilitiesQueryManager: QueryManager<null, Capabilities>;
+
+  static getAxios() {
+    return axios;
+  }
 
   static shownNotifications() {
     AppToaster.show({
