+++ web-console/src/druid-models/ingestion-spec.tsx
@@ -19,19 +19,16 @@
 import { Code } from '@blueprintjs/core';
 import React from 'react';
 
-import { Field } from '../components/auto-form/auto-form';
-import { ExternalLink } from '../components/external-link/external-link';
-import { TransformSpec } from '../druid-models';
+import { ExternalLink, Field } from '../components';
 import { getLink } from '../links';
+import { deepDelete, deepGet, deepMove, deepSet, oneOf } from '../utils';
 
-import {
-  BASIC_TIME_FORMATS,
-  DATE_ONLY_TIME_FORMATS,
-  DATETIME_TIME_FORMATS,
-  OTHER_TIME_FORMATS,
-} from './druid-time';
-import { oneOf } from './general';
-import { deepDelete, deepGet, deepMove, deepSet } from './object-change';
+import { DimensionsSpec } from './dimension-spec';
+import { InputFormat, issueWithInputFormat } from './input-format';
+import { InputSource, issueWithInputSource } from './input-source';
+import { MetricSpec } from './metric-spec';
+import { PLACEHOLDER_TIMESTAMP_SPEC, TimestampSpec } from './timestamp-spec';
+import { TransformSpec } from './transform-spec';
 
 export const MAX_INLINE_DATA_LENGTH = 65536;
 
@@ -223,18 +220,6 @@ export interface DataSchema {
   metricsSpec?: MetricSpec[];
 }
 
-export interface InputFormat {
-  type: string;
-  findColumnsFromHeader?: boolean;
-  skipHeaderRows?: number;
-  columns?: string[];
-  listDelimiter?: string;
-  pattern?: string;
-  function?: string;
-  flattenSpec?: FlattenSpec;
-  keepNullColumns?: boolean;
-}
-
 export type DimensionMode = 'specific' | 'auto-detect';
 
 export function getDimensionMode(spec: IngestionSpec): DimensionMode {
@@ -308,611 +293,12 @@ export function cleanSpec(spec: IngestionSpec): IngestionSpec {
   };
 }
 
-const INPUT_FORMAT_FORM_FIELDS: Field<InputFormat>[] = [
-  {
-    name: 'type',
-    label: 'Input format',
-    type: 'string',
-    suggestions: ['json', 'csv', 'tsv', 'regex', 'parquet', 'orc', 'avro_ocf'],
-    required: true,
-    info: (
-      <>
-        <p>The parser used to parse the data.</p>
-        <p>
-          For more information see{' '}
-          <ExternalLink href={`${getLink('DOCS')}/ingestion/data-formats.html`}>
-            the documentation
-          </ExternalLink>
-          .
-        </p>
-      </>
-    ),
-  },
-  {
-    name: 'pattern',
-    type: 'string',
-    required: true,
-    defined: (p: InputFormat) => p.type === 'regex',
-  },
-  {
-    name: 'function',
-    type: 'string',
-    required: true,
-    defined: (p: InputFormat) => p.type === 'javascript',
-  },
-  {
-    name: 'findColumnsFromHeader',
-    type: 'boolean',
-    required: true,
-    defined: (p: InputFormat) => oneOf(p.type, 'csv', 'tsv'),
-  },
-  {
-    name: 'skipHeaderRows',
-    type: 'number',
-    defaultValue: 0,
-    defined: (p: InputFormat) => oneOf(p.type, 'csv', 'tsv'),
-    min: 0,
-    info: (
-      <>
-        If both skipHeaderRows and hasHeaderRow options are set, skipHeaderRows is first applied.
-        For example, if you set skipHeaderRows to 2 and hasHeaderRow to true, Druid will skip the
-        first two lines and then extract column information from the third line.
-      </>
-    ),
-  },
-  {
-    name: 'columns',
-    type: 'string-array',
-    required: (p: InputFormat) =>
-      (oneOf(p.type, 'csv', 'tsv') && !p.findColumnsFromHeader) || p.type === 'regex',
-    defined: (p: InputFormat) =>
-      (oneOf(p.type, 'csv', 'tsv') && !p.findColumnsFromHeader) || p.type === 'regex',
-  },
-  {
-    name: 'delimiter',
-    type: 'string',
-    defaultValue: '\t',
-    defined: (p: InputFormat) => p.type === 'tsv',
-    info: <>A custom delimiter for data values.</>,
-  },
-  {
-    name: 'listDelimiter',
-    type: 'string',
-    defined: (p: InputFormat) => oneOf(p.type, 'csv', 'tsv', 'regex'),
-    info: <>A custom delimiter for multi-value dimensions.</>,
-  },
-  {
-    name: 'binaryAsString',
-    type: 'boolean',
-    defaultValue: false,
-    defined: (p: InputFormat) => oneOf(p.type, 'parquet', 'orc', 'avro_ocf'),
-    info: (
-      <>
-        Specifies if the bytes parquet column which is not logically marked as a string or enum type
-        should be treated as a UTF-8 encoded string.
-      </>
-    ),
-  },
-];
-
-export function getInputFormatFormFields() {
-  return INPUT_FORMAT_FORM_FIELDS;
-}
-
-export function issueWithInputFormat(inputFormat: InputFormat | undefined): string | undefined {
-  if (!inputFormat) return 'no input format';
-  if (!inputFormat.type) return 'missing a type';
-  switch (inputFormat.type) {
-    case 'regex':
-      if (!inputFormat.pattern) return "must have a 'pattern'";
-      break;
-
-    case 'javascript':
-      if (!inputFormat['function']) return "must have a 'function'";
-      break;
-  }
-  return;
-}
-
-export function inputFormatCanFlatten(inputFormat: InputFormat): boolean {
-  const inputFormatType = inputFormat.type;
-  return (
-    inputFormatType === 'json' ||
-    inputFormatType === 'parquet' ||
-    inputFormatType === 'orc' ||
-    inputFormatType === 'avro_ocf'
-  );
-}
-
-export interface TimestampSpec {
-  column?: string;
-  format?: string;
-  missingValue?: string;
-}
-
-export function getTimestampSpecColumn(timestampSpec: TimestampSpec) {
-  // https://github.com/apache/druid/blob/master/core/src/main/java/org/apache/druid/data/input/impl/TimestampSpec.java#L44
-  return timestampSpec.column || 'timestamp';
-}
-
-const NO_SUCH_COLUMN = '!!!_no_such_column_!!!';
-
-const DUMMY_TIMESTAMP_SPEC: TimestampSpec = {
-  column: NO_SUCH_COLUMN,
-  missingValue: '1970-01-01T00:00:00Z',
-};
-
-export function getDummyTimestampSpec() {
-  return DUMMY_TIMESTAMP_SPEC;
-}
-
-const CONSTANT_TIMESTAMP_SPEC: TimestampSpec = {
-  column: NO_SUCH_COLUMN,
-  missingValue: '2010-01-01T00:00:00Z',
-};
-
-export function getConstantTimestampSpec() {
-  return CONSTANT_TIMESTAMP_SPEC;
-}
-
-export function isColumnTimestampSpec(timestampSpec: TimestampSpec) {
-  return (deepGet(timestampSpec, 'column') || 'timestamp') !== NO_SUCH_COLUMN;
-}
-
-const TIMESTAMP_SPEC_FORM_FIELDS: Field<TimestampSpec>[] = [
-  {
-    name: 'column',
-    type: 'string',
-    defaultValue: 'timestamp',
-  },
-  {
-    name: 'format',
-    type: 'string',
-    defaultValue: 'auto',
-    suggestions: [
-      ...BASIC_TIME_FORMATS,
-      {
-        group: 'Date and time formats',
-        suggestions: DATETIME_TIME_FORMATS,
-      },
-      {
-        group: 'Date only formats',
-        suggestions: DATE_ONLY_TIME_FORMATS,
-      },
-      {
-        group: 'Other time formats',
-        suggestions: OTHER_TIME_FORMATS,
-      },
-    ],
-    defined: (timestampSpec: TimestampSpec) => isColumnTimestampSpec(timestampSpec),
-    info: (
-      <p>
-        Please specify your timestamp format by using the suggestions menu or typing in a{' '}
-        <ExternalLink href="https://docs.oracle.com/javase/8/docs/api/java/time/format/DateTimeFormatter.html">
-          format string
-        </ExternalLink>
-        .
-      </p>
-    ),
-  },
-  {
-    name: 'missingValue',
-    type: 'string',
-    placeholder: '(optional)',
-    info: <p>This value will be used if the specified column can not be found.</p>,
-  },
-];
-
-const CONSTANT_TIMESTAMP_SPEC_FORM_FIELDS: Field<TimestampSpec>[] = [
-  {
-    name: 'missingValue',
-    label: 'Constant value',
-    type: 'string',
-    info: <p>The dummy value that will be used as the timestamp.</p>,
-  },
-];
-
-export function getTimestampSpecFormFields(timestampSpec: TimestampSpec) {
-  if (isColumnTimestampSpec(timestampSpec)) {
-    return TIMESTAMP_SPEC_FORM_FIELDS;
-  } else {
-    return CONSTANT_TIMESTAMP_SPEC_FORM_FIELDS;
-  }
-}
-
-export function issueWithTimestampSpec(
-  timestampSpec: TimestampSpec | undefined,
-): string | undefined {
-  if (!timestampSpec) return 'no spec';
-  if (!timestampSpec.column && !timestampSpec.missingValue) return 'timestamp spec is blank';
-  return;
-}
-
-export interface DimensionsSpec {
-  dimensions?: (string | DimensionSpec)[];
-  dimensionExclusions?: string[];
-  spatialDimensions?: any[];
-}
-
-export interface DimensionSpec {
-  type: string;
-  name: string;
-  createBitmapIndex?: boolean;
-}
-
-const DIMENSION_SPEC_FORM_FIELDS: Field<DimensionSpec>[] = [
-  {
-    name: 'name',
-    type: 'string',
-  },
-  {
-    name: 'type',
-    type: 'string',
-    suggestions: ['string', 'long', 'float', 'double'],
-  },
-  {
-    name: 'createBitmapIndex',
-    type: 'boolean',
-    defaultValue: true,
-    defined: (dimensionSpec: DimensionSpec) => dimensionSpec.type === 'string',
-  },
-];
-
-export function getDimensionSpecFormFields() {
-  return DIMENSION_SPEC_FORM_FIELDS;
-}
-
-export function getDimensionSpecName(dimensionSpec: string | DimensionSpec): string {
-  return typeof dimensionSpec === 'string' ? dimensionSpec : dimensionSpec.name;
-}
-
-export function getDimensionSpecType(dimensionSpec: string | DimensionSpec): string {
-  return typeof dimensionSpec === 'string' ? 'string' : dimensionSpec.type;
-}
-
-export function inflateDimensionSpec(dimensionSpec: string | DimensionSpec): DimensionSpec {
-  return typeof dimensionSpec === 'string'
-    ? { name: dimensionSpec, type: 'string' }
-    : dimensionSpec;
-}
-
-export interface FlattenSpec {
-  useFieldDiscovery?: boolean;
-  fields?: FlattenField[];
-}
-
-export interface FlattenField {
-  name: string;
-  type: string;
-  expr: string;
-}
-
-const FLATTEN_FIELD_FORM_FIELDS: Field<FlattenField>[] = [
-  {
-    name: 'name',
-    type: 'string',
-    placeholder: 'column_name',
-    required: true,
-  },
-  {
-    name: 'type',
-    type: 'string',
-    suggestions: ['path', 'jq', 'root'],
-    required: true,
-  },
-  {
-    name: 'expr',
-    type: 'string',
-    placeholder: '$.thing',
-    defined: (flattenField: FlattenField) => oneOf(flattenField.type, 'path', 'jq'),
-    required: true,
-    info: (
-      <>
-        Specify a flatten{' '}
-        <ExternalLink href={`${getLink('DOCS')}/ingestion/flatten-json`}>expression</ExternalLink>.
-      </>
-    ),
-  },
-];
-
-export function getFlattenFieldFormFields() {
-  return FLATTEN_FIELD_FORM_FIELDS;
-}
-
 export interface GranularitySpec {
   type?: string;
   queryGranularity?: string;
   segmentGranularity?: string;
   rollup?: boolean;
   intervals?: string | string[];
-}
-
-export interface MetricSpec {
-  type: string;
-  name?: string;
-  fieldName?: string;
-  maxStringBytes?: number;
-  filterNullValues?: boolean;
-  fieldNames?: string[];
-  fnAggregate?: string;
-  fnCombine?: string;
-  fnReset?: string;
-  fields?: string[];
-  byRow?: boolean;
-  round?: boolean;
-  isInputHyperUnique?: boolean;
-  filter?: any;
-  aggregator?: MetricSpec;
-}
-
-const METRIC_SPEC_FORM_FIELDS: Field<MetricSpec>[] = [
-  {
-    name: 'name',
-    type: 'string',
-    info: <>The metric name as it will appear in Druid.</>,
-  },
-  {
-    name: 'type',
-    type: 'string',
-    suggestions: [
-      'count',
-      {
-        group: 'sum',
-        suggestions: ['longSum', 'doubleSum', 'floatSum'],
-      },
-      {
-        group: 'min',
-        suggestions: ['longMin', 'doubleMin', 'floatMin'],
-      },
-      {
-        group: 'max',
-        suggestions: ['longMax', 'doubleMax', 'floatMax'],
-      },
-      {
-        group: 'first',
-        suggestions: ['longFirst', 'doubleFirst', 'floatFirst'],
-      },
-      {
-        group: 'last',
-        suggestions: ['longLast', 'doubleLast', 'floatLast'],
-      },
-      'thetaSketch',
-      {
-        group: 'HLLSketch',
-        suggestions: ['HLLSketchBuild', 'HLLSketchMerge'],
-      },
-      'quantilesDoublesSketch',
-      'momentSketch',
-      'fixedBucketsHistogram',
-      'hyperUnique',
-      'filtered',
-    ],
-    info: <>The aggregation function to apply.</>,
-  },
-  {
-    name: 'fieldName',
-    type: 'string',
-    defined: m => m.type !== 'filtered',
-    info: <>The column name for the aggregator to operate on.</>,
-  },
-  {
-    name: 'maxStringBytes',
-    type: 'number',
-    defaultValue: 1024,
-    defined: m => {
-      return oneOf(m.type, 'stringFirst', 'stringLast');
-    },
-  },
-  {
-    name: 'filterNullValues',
-    type: 'boolean',
-    defaultValue: false,
-    defined: m => {
-      return oneOf(m.type, 'stringFirst', 'stringLast');
-    },
-  },
-  // filtered
-  {
-    name: 'filter',
-    type: 'json',
-    defined: m => m.type === 'filtered',
-  },
-  {
-    name: 'aggregator',
-    type: 'json',
-    defined: m => m.type === 'filtered',
-  },
-  // thetaSketch
-  {
-    name: 'size',
-    type: 'number',
-    defined: m => m.type === 'thetaSketch',
-    defaultValue: 16384,
-    info: (
-      <>
-        <p>
-          Must be a power of 2. Internally, size refers to the maximum number of entries sketch
-          object will retain. Higher size means higher accuracy but more space to store sketches.
-          Note that after you index with a particular size, druid will persist sketch in segments
-          and you will use size greater or equal to that at query time.
-        </p>
-        <p>
-          See the{' '}
-          <ExternalLink href="https://datasketches.apache.org/docs/Theta/ThetaSize.html">
-            DataSketches site
-          </ExternalLink>{' '}
-          for details.
-        </p>
-        <p>In general, We recommend just sticking to default size.</p>
-      </>
-    ),
-  },
-  {
-    name: 'isInputThetaSketch',
-    type: 'boolean',
-    defined: m => m.type === 'thetaSketch',
-    defaultValue: false,
-    info: (
-      <>
-        This should only be used at indexing time if your input data contains theta sketch objects.
-        This would be the case if you use datasketches library outside of Druid, say with Pig/Hive,
-        to produce the data that you are ingesting into Druid
-      </>
-    ),
-  },
-  // HLLSketchBuild & HLLSketchMerge
-  {
-    name: 'lgK',
-    type: 'number',
-    defined: m => oneOf(m.type, 'HLLSketchBuild', 'HLLSketchMerge'),
-    defaultValue: 12,
-    info: (
-      <>
-        <p>
-          log2 of K that is the number of buckets in the sketch, parameter that controls the size
-          and the accuracy.
-        </p>
-        <p>Must be between 4 to 21 inclusively.</p>
-      </>
-    ),
-  },
-  {
-    name: 'tgtHllType',
-    type: 'string',
-    defined: m => oneOf(m.type, 'HLLSketchBuild', 'HLLSketchMerge'),
-    defaultValue: 'HLL_4',
-    suggestions: ['HLL_4', 'HLL_6', 'HLL_8'],
-    info: (
-      <>
-        The type of the target HLL sketch. Must be <Code>HLL_4</Code>, <Code>HLL_6</Code>, or{' '}
-        <Code>HLL_8</Code>.
-      </>
-    ),
-  },
-  // quantilesDoublesSketch
-  {
-    name: 'k',
-    type: 'number',
-    defined: m => m.type === 'quantilesDoublesSketch',
-    defaultValue: 128,
-    info: (
-      <>
-        <p>
-          Parameter that determines the accuracy and size of the sketch. Higher k means higher
-          accuracy but more space to store sketches.
-        </p>
-        <p>
-          Must be a power of 2 from 2 to 32768. See the{' '}
-          <ExternalLink href="https://datasketches.apache.org/docs/Quantiles/QuantilesAccuracy.html">
-            Quantiles Accuracy
-          </ExternalLink>{' '}
-          for details.
-        </p>
-      </>
-    ),
-  },
-  // momentSketch
-  {
-    name: 'k',
-    type: 'number',
-    defined: m => m.type === 'momentSketch',
-    required: true,
-    info: (
-      <>
-        Parameter that determines the accuracy and size of the sketch. Higher k means higher
-        accuracy but more space to store sketches. Usable range is generally [3,15]
-      </>
-    ),
-  },
-  {
-    name: 'compress',
-    type: 'boolean',
-    defined: m => m.type === 'momentSketch',
-    defaultValue: true,
-    info: (
-      <>
-        Flag for whether the aggregator compresses numeric values using arcsinh. Can improve
-        robustness to skewed and long-tailed distributions, but reduces accuracy slightly on more
-        uniform distributions.
-      </>
-    ),
-  },
-  // fixedBucketsHistogram
-  {
-    name: 'lowerLimit',
-    type: 'number',
-    defined: m => m.type === 'fixedBucketsHistogram',
-    required: true,
-    info: <>Lower limit of the histogram.</>,
-  },
-  {
-    name: 'upperLimit',
-    type: 'number',
-    defined: m => m.type === 'fixedBucketsHistogram',
-    required: true,
-    info: <>Upper limit of the histogram.</>,
-  },
-  {
-    name: 'numBuckets',
-    type: 'number',
-    defined: m => m.type === 'fixedBucketsHistogram',
-    defaultValue: 10,
-    required: true,
-    info: (
-      <>
-        Number of buckets for the histogram. The range <Code>[lowerLimit, upperLimit]</Code> will be
-        divided into <Code>numBuckets</Code> intervals of equal size.
-      </>
-    ),
-  },
-  {
-    name: 'outlierHandlingMode',
-    type: 'string',
-    defined: m => m.type === 'fixedBucketsHistogram',
-    required: true,
-    suggestions: ['ignore', 'overflow', 'clip'],
-    info: (
-      <>
-        <p>
-          Specifies how values outside of <Code>[lowerLimit, upperLimit]</Code> will be handled.
-        </p>
-        <p>
-          Supported modes are <Code>ignore</Code>, <Code>overflow</Code>, and <Code>clip</Code>. See
-          <ExternalLink
-            href={`${getLink(
-              'DOCS',
-            )}/development/extensions-core/approximate-histograms.html#outlier-handling-modes`}
-          >
-            outlier handling modes
-          </ExternalLink>{' '}
-          for more details.
-        </p>
-      </>
-    ),
-  },
-  // hyperUnique
-  {
-    name: 'isInputHyperUnique',
-    type: 'boolean',
-    defined: m => m.type === 'hyperUnique',
-    defaultValue: false,
-    info: (
-      <>
-        This can be set to true to index precomputed HLL (Base64 encoded output from druid-hll is
-        expected).
-      </>
-    ),
-  },
-];
-
-export function getMetricSpecFormFields() {
-  return METRIC_SPEC_FORM_FIELDS;
-}
-
-export function getMetricSpecName(metricSpec: MetricSpec): string {
-  return (
-    metricSpec.name || (metricSpec.aggregator ? getMetricSpecName(metricSpec.aggregator) : '?')
-  );
 }
 
 // --------------
@@ -940,29 +326,6 @@ export function invalidIoConfig(ioConfig: IoConfig): boolean {
     (ioConfig.type === 'kafka' && ioConfig.useEarliestOffset == null) ||
     (ioConfig.type === 'kinesis' && ioConfig.useEarliestSequenceNumber == null)
   );
-}
-
-export interface InputSource {
-  type: string;
-  baseDir?: string;
-  filter?: any;
-  uris?: string[];
-  prefixes?: string[];
-  objects?: { bucket: string; path: string }[];
-  fetchTimeout?: number;
-
-  // druid
-  dataSource?: string;
-  interval?: string;
-  dimensions?: string[];
-  metrics?: string[];
-  maxInputSegmentBytesPerTask?: number;
-
-  // inline
-  data?: string;
-
-  // hdfs
-  paths?: string;
 }
 
 export function getIoConfigFormFields(ingestionComboType: IngestionComboType): Field<IoConfig>[] {
@@ -1538,55 +901,6 @@ export function getIoConfigFormFields(ingestionComboType: IngestionComboType): F
   }
 
   throw new Error(`unknown input type ${ingestionComboType}`);
-}
-
-function nonEmptyArray(a: any) {
-  return Array.isArray(a) && Boolean(a.length);
-}
-
-function issueWithInputSource(inputSource: InputSource | undefined): string | undefined {
-  if (!inputSource) return 'does not exist';
-  if (!inputSource.type) return 'missing a type';
-  switch (inputSource.type) {
-    case 'local':
-      if (!inputSource.baseDir) return `must have a 'baseDir'`;
-      if (!inputSource.filter) return `must have a 'filter'`;
-      break;
-
-    case 'http':
-      if (!nonEmptyArray(inputSource.uris)) {
-        return 'must have at least one uri';
-      }
-      break;
-
-    case 'druid':
-      if (!inputSource.dataSource) return `must have a 'dataSource'`;
-      if (!inputSource.interval) return `must have an 'interval'`;
-      break;
-
-    case 'inline':
-      if (!inputSource.data) return `must have 'data'`;
-      break;
-
-    case 's3':
-    case 'azure':
-    case 'google':
-      if (
-        !nonEmptyArray(inputSource.uris) &&
-        !nonEmptyArray(inputSource.prefixes) &&
-        !nonEmptyArray(inputSource.objects)
-      ) {
-        return 'must have at least one uri or prefix or object';
-      }
-      break;
-
-    case 'hdfs':
-      if (!inputSource.paths) {
-        return 'must have paths';
-      }
-      break;
-  }
-  return;
 }
 
 export function issueWithIoConfig(
@@ -2623,7 +1937,7 @@ export function updateIngestionType(
   }
 
   if (!deepGet(spec, 'spec.dataSchema.timestampSpec')) {
-    newSpec = deepSet(newSpec, 'spec.dataSchema.timestampSpec', getDummyTimestampSpec());
+    newSpec = deepSet(newSpec, 'spec.dataSchema.timestampSpec', PLACEHOLDER_TIMESTAMP_SPEC);
   }
 
   if (!deepGet(spec, 'spec.dataSchema.dimensionsSpec')) {
