+++ web-console/src/entry.scss
@@ -16,13 +16,13 @@
  * limitations under the License.
  */
 
-@import '../node_modules/normalize.css/normalize';
+@import '~normalize.css/normalize';
 @import '~fontsource-open-sans/index.css';
 @import './blueprint-overrides';
 @import '~@blueprintjs/core/src/blueprint';
 @import '~@blueprintjs/datetime/src/blueprint-datetime';
+@import '~react-splitter-layout/lib/index';
 @import '../lib/react-table';
-@import '../node_modules/react-splitter-layout/lib/index.css';
 
 html,
 body {
